added web rendering, still need to impl input

+++ internal/rendering/shader_sources/vertex.glsl
@@ -1,7 +1,7 @@
 <version>
 layout(location = 0) in vec2 aPos;
 // color.w/a/q is layer
-layout(location = 1) in ivec4 aColor;
+layout(location = 1) in uvec4 aColor;
 // layouts from channels
 <attributes>
 
@@ -24,10 +24,11 @@ void main() {
 
     gl_Position = vec4(
         // position
-        (<xAxis> * aPos.x + <yAxis> * -aPos.y + <pos>) / vec2(screenSize) * vec2(2, -2) + vec2(-1, 1),
-        0, 1.0
+        (<xAxis> * aPos.x + <yAxis> * -aPos.y + <pos>) / vec2(screenSize) * vec2(2, -2) + vec2(-1, 1), 
+        0.0, 1.0
+        //0.0, 0.0, 0.0, 1.0
     );
 
-    vertexColor = vec4(vec3(<color>)/256.0, 1);
-    layer = <layer>*uint(256) + uint(aColor.a+1);
+    vertexColor = vec4(vec3(<color>)/256.0, 1.0);
+    layer = <layer>*256u + aColor.a+1u;
 }
